requirements.txt
# Core dependencies
ccxt==4.1.76
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
joblib==1.3.2
talib-binary==0.4.28
python-dotenv==1.0.0

# Web and async
aiohttp==3.8.5
requests==2.31.0

# Optional for ML (if training locally)
tensorflow==2.13.0
optuna==3.3.0